Move useShippingRatesWithSavedSuggestions files, to be used in SetupFreeListings.
